chapter4/agent-with-event-trigger/requirements.txt
openai>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
mcp>=1.0.0